Pega eventos da base de tratativas do mariadb e sobe pra base de Analytics correspondente

Pega eventos da base de tratativas do mariadb e sobe pra base de Analytics correspondente

consolida_base_tratativa_mariadb_nova(
  conexao_db_original,
  conexao_db_tratativas,
  db_selecionado,
  usar_campo_empresa = TRUE,
  data_inicial = NULL,
  data_final = NULL
)

Arguments

conexao_db_original
Conexão com o db original de Analytics

conexao_db_tratativas
Conexão com o db de tratativas

db_selecionado
Nome do db selecionado (pra saber quais operações da tratativa ele tem que pegar)

usar_campo_empresa
Se TRUE, busca o nome da empresa na coluna Nome_database da planilha Controle. Se FALSE, busca pelo padrão do nome da operação.

data_inicial
Se NULL, usa '2018-01-01'

data_final
Se NULL, usa o dia de ontem

Value

Retorna nada fi