Envia emails usando o Outlook

Monte emails usando o R e envie com o Outlook, com anexos, múltiplos contatos, etc. O email pronto será exibido e bastará apertar enviar. Só funciona usando R 3.6 (e não o 4).

Usage

email_outlook(
  para = "",
  cc = "",
  bcc = "",
  assunto = "",
  texto_email = "",
  assinatura = "",
  anexos = "",
  exibir_email = TRUE,
  enviar_email = FALSE,
  usar_utf8 = TRUE,
  anexo_utf8 = TRUE
)

Arguments

para

Um vetor com emails (o que se escreve no 'para').

cc

Um vetor com emails em cópia 'cc'.

bcc

Um vetor com emails para cópia oculta.

assunto

O assunto do email.

texto_email

O texto do email (em html).

assinatura

A assinatura do email (em html).

anexos

Um vetor com o endereço dos anexos (o caminho pode estar relativo ao seu working directory).

exibir_email

Se TRUE (padrão), exibe o email montado no Outlook (e voce envia apos checar).

enviar_email

Se TRUE (o padrão é FALSE), envia o email logo após montar.

usar_utf8

Se FALSE (o padrão é TRUE), converte o assunto e o texto pro formato native.

Details

O pacote RDCOMclient é meio chato de instalar. Se ao usar essa função ele der erro de instalação, provavelmente terá que apagar um arquivo chamado LOCK-não-sei-o-que (conforme escrito no erro que der) e tentar usar a função de novo. É preciso instalar o Rtools35.